third_party/WebKit/LayoutTests/editing/execCommand/insert_list_infinite_loop2.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
  '<div></div>',
  '<table contenteditable="true><tbody><tr><td>',
    '^<ol>',
      '<li>AAA<br><br></li><li><br></li><li>BBB<br>CCC</li>',
    '</ol>|',
  '</td></tr></tbody></table>',
  'DDD',
  '<div style="-webkit-user-select: none;"></div>',
  '<div style="display: inline-block; border:solid;"></div>'
  ],
  'insertOrderedList',
  []),
  'This tests InsertList inside an empty div with no placeholder');
</script>
